+++ sdk/appcenter-auth/build.gradle
