+++ sdk/appcenter-auth/proguard-rules.pro
