+++ sdk/appcenter-auth/src/test/java/com/microsoft/appcenter/auth/AbstractIdentityTest.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-package com.microsoft.appcenter.identity;
+package com.microsoft.appcenter.auth;
 
 import android.content.Context;
 import android.os.SystemClock;
