+++ settings.gradle
@@ -9,7 +9,7 @@ include ':sdk:appcenter'
 include ':sdk:appcenter-crashes'
 include ':sdk:appcenter-analytics'
 include ':sdk:appcenter-distribute'
-include ':sdk:appcenter-identity'
+include ':sdk:appcenter-auth'
 include ':sdk:appcenter-storage'
 include ':sdk:appcenter-push'
 include ':sdk:appcenter-rum'
